requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
pytest>=6.2.4
requests>=2.0,<3.0
psycopg2-binary>=2.9.0,<3.0.0
